+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored `WsiDicomWebClient.__init__()` to take a `DICOMwebClient` instance, and moved the old `__init__()` method into `WsiDicomWebClient.create_client()`.
 - Make some arguments optional for `WsiDicomWebClient.create_client()`.
+- Removed `WsiDicomFileClient` since it is no longer needed.
 
 ## [0.12.0] - 2023-10-04